requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
coverage==7.6.1
django==5.1.1
django-widget-tweaks==1.5.0
exceptiongroup==1.2.2
faker==28.4.1
idna==3.9
iniconfig==2.0.0
model-bakery==1.19.5
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
typing-extensions==4.12.2
urllib3==2.2.3
